requirements.txt
typer[all]
google-auth-httplib2
google-api-python-client
google-auth-oauthlib
pytest
pytest-cov
platformdirs
black
